Fitbit Data Analyzer/On Screen HR/requirements.txt
PyQt
qtawesome
keyboard